backend/requirements.txt
Django==5.1.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
celery==5.3.6
redis==5.0.1
Pillow==10.2.0
python-dotenv==1.0.1
whitenoise==6.6.0
gunicorn==21.2.0 